[KAN-86] rename: 유저관련 서버액션 폴더 이름 변경
