+++ src/app/api/actions/user/putProfileData.ts
@@ -3,11 +3,18 @@
 import { getCookie } from '@/actions/auth/cookie/cookie';
 import { UserData } from '@/types/client.type';
 import { revalidatePath } from 'next/cache';
+import { redirect } from 'next/navigation';
 
 export const putProfileData = async (
   formData: FormData,
 ): Promise<UserData | null> => {
   const token = await getCookie('token');
+
+  if (!token) {
+    alert('로그인 세션이 만료되었습니다. 다시 로그인 해주세요.');
+    redirect('/signin');
+    return null;
+  }
 
   try {
     const response = await fetch(
@@ -26,7 +33,7 @@ export const putProfileData = async (
     }
 
     const result: UserData = await response.json();
-    revalidatePath('/');
+    revalidatePath('/auths/user');
 
     return result; // 업데이트된 사용자 데이터를 반환
   } catch (error) {
