+++ src/app/components/UserProfileLayout/UserProfileLayout.tsx
@@ -4,12 +4,12 @@ import { Profile } from '@/public/images';
 import ProfileEditModal from '../Modal/ProfileEditModal';
 import { useState } from 'react';
 import onChangeProfileImage from './onChangeProfileImage';
-import { putProfileData } from '@/app/api/actions/mypage/putProfileData';
 import { useProfileState } from './useProfileState';
 import Image from 'next/image';
 import UserProfileHeader from './UserProfileHeader';
 import UserInfo from './UserInfo';
 import { UserData } from '@/types/client.type';
+import { putProfileData } from '@/app/api/actions/user/putProfileData';
 
 interface MyGatheringListProps {
   user: UserData | null;
